refactor(ai-providers): show provider name in on-device attribution and share the System Settings opener
